# [PERF] stock: speedup get_orderpoint_action