@robodoo 2025-03-25
@auma-odoo retry makes no sense when the PR is not in error.