@bpcreate 2025-04-01
The performance increase for get_orderpoint_action is indeed significant, however the problem still remains when the To Order filter (qty_to_order > 0) domain is queried. Are there any open PRs that address this?