@auma-odoo 2025-04-17
@bpcreate Sorry, I didn't see your message before. From what I know, this filter is tricky to speed up since it need to recompute most `qty_to_order` because they depend on so many variable. We have some fix that should speed that up but it won't make it blazingly fast :sweat:. The fix are: #202365 in 17.0 and #195608 but for master only.