@bpcreate 2025-04-17
Thanks for your reply, What was the breaking change? In v17 this process
takes less than 5 seconds for 15,000 orderpoints.

On Thu, Apr 17, 2025, 8:37 AM Aurélien MARIE ***@***.***>
wrote:

> @bpcreate <https://github.com/bpcreate> Sorry, I didn't see your message
> before. From what I know, this filter is tricky to speed up since it need
> to recompute most qty_to_order because they depend on so many variable.
> We have some fix that should speed that up but it won't make it blazingly
> fast 😓. The fix are: #202365 <https://github.com/odoo/odoo/pull/202365>
> in 17.0 and #195608 <https://github.com/odoo/odoo/pull/195608> but for
> master only.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/odoo/odoo/pull/203276#issuecomment-2813151361>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/BD3NE35TKPQTBUWTTBKHM7T2Z64A5AVCNFSM6AAAAABZXTDA2OVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDQMJTGE2TCMZWGE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
> *auma-odoo* left a comment (odoo/odoo#203276)
> <https://github.com/odoo/odoo/pull/203276#issuecomment-2813151361>
>
> @bpcreate <https://github.com/bpcreate> Sorry, I didn't see your message
> before. From what I know, this filter is tricky to speed up since it need
> to recompute most qty_to_order because they depend on so many variable.
> We have some fix that should speed that up but it won't make it blazingly
> fast 😓. The fix are: #202365 <https://github.com/odoo/odoo/pull/202365>
> in 17.0 and #195608 <https://github.com/odoo/odoo/pull/195608> but for
> master only.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/odoo/odoo/pull/203276#issuecomment-2813151361>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/BD3NE35TKPQTBUWTTBKHM7T2Z64A5AVCNFSM6AAAAABZXTDA2OVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDQMJTGE2TCMZWGE>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>