@auma-odoo 2025-04-22
The change was caused by this PR #159432. The purpose is to remove the impact on writing from other model and to only impact the read (the replenishment view for example)